Export data UI (#8)
* added confirm box to view media + export data ui

* edited view media + export data ui

* cleaned up code

* updated design

+++ App/Views/ViewMediaView.swift
@@ -13,14 +13,11 @@ struct MediaItemView: View {
     let item: Media
 
     var body: some View {
-        VStack() {
-            Image(uiImage: UIImage(data: item.data!)!)
-                    .resizable()
-                    .scaledToFit()
-                    .padding(.bottom, 10)
-            Text(item.name!)
-                .padding(.bottom, 10)
-        }
+        Image(uiImage: UIImage(data: item.data!)!)
+                .resizable()
+                .scaledToFit()
+                .cornerRadius(15)
+                .frame(width: 970)
     }
 }
 
@@ -32,19 +29,31 @@ struct ViewMediaView: View {
     
     @Environment(\.managedObjectContext) private var viewContext
     
-    private var gridItems = [GridItem(.flexible()), GridItem(.flexible()), GridItem(.flexible())]
-    
     @FetchRequest(entity: Media.entity(), sortDescriptors: [])
         var mediaData: FetchedResults<Media>
+    
+    private var gridItems = [GridItem(.flexible()), GridItem(.flexible()), GridItem(.flexible())]
     
     var body: some View {
         LazyVGrid(columns: gridItems, spacing: 10) {
             ForEach(mediaData.indices, id: \.self) { i in
                 let item =  mediaData[i]
-                let mediaName = item.name
-                    ZStack {
-                        NavigationLink(destination: MediaItemView(item: item)) {
-                            VStack(spacing: 10) {
+                VStack {
+                    ZStack (alignment: .topTrailing) {
+                        VStack(spacing: 10) {
+                            Text("x")
+                                .foregroundColor(Color.white)
+                                .frame(width: 40, height: 40)
+                                .background(Color.clear)
+                                .offset(x: 15, y: 45)
+                            Button(action : {
+                                let alertHC = UIHostingController(rootView: MediaItemView(item: item))
+                                alertHC.view.backgroundColor = .clear
+                                alertHC.preferredContentSize = CGSize(width: 1000, height: 1000)
+                                alertHC.modalPresentationStyle = UIModalPresentationStyle.formSheet
+
+                                UIApplication.shared.windows[0].rootViewController?.present(alertHC, animated: true)
+                            }) {
                                 Image(uiImage: UIImage(data: item.data!)!)
                                     .resizable()
                                     .scaledToFit()
@@ -52,58 +61,82 @@ struct ViewMediaView: View {
                                     .cornerRadius(10)
                                     .padding(.top, 10)
                                     .padding(.bottom, 5)
-                                Button(action : {
-                                    let alertHC = UIHostingController(rootView: AlertView(item: item).environment(\.managedObjectContext, viewContext))
-
-                                    alertHC.preferredContentSize = CGSize(width: 300, height: 175)
-                                    alertHC.modalPresentationStyle = UIModalPresentationStyle.formSheet
-
-                                    UIApplication.shared.windows[0].rootViewController?.present(alertHC, animated: true)
-                                })
-                                {
-                                    Text(mediaName!)
-                                    .minimumScaleFactor(0.5)
-                                    .padding(.bottom, 10)
-                                }
                             }
                         }
                         Button(action : {
-                            let _ = viewContext.delete(item)
-                            try? viewContext.save()
+                            let alertHC = UIHostingController(rootView: ConfirmView(item: item).environment(\.managedObjectContext, viewContext))
+                            alertHC.preferredContentSize = CGSize(width: 320, height: 115)
+                            alertHC.modalPresentationStyle = UIModalPresentationStyle.formSheet
+                            UIApplication.shared.windows[0].rootViewController?.present(alertHC, animated: true)
                         })
                         {
                             Text("x")
+                                .font(.system(size: 20.0))
                                 .foregroundColor(.white)
-                                .frame(width:40,height:40)
+                                .frame(width: 40, height: 40)
                                 .background(Color.black)
                                 .cornerRadius(100)
-                        }.offset(x:130,y:-90)
+                        }.offset(x: 15, y: 45)
+                    }.padding(.top, -10)
+                    Button(action : {
+                        let alertHC = UIHostingController(rootView: AlertView(item: item).environment(\.managedObjectContext, viewContext))
+                        alertHC.preferredContentSize = CGSize(width: 500, height: 175)
+                        alertHC.modalPresentationStyle = UIModalPresentationStyle.formSheet
+                        UIApplication.shared.windows[0].rootViewController?.present(alertHC, animated: true)
+                    })
+                    {
+                        Text(item.name!).padding(10).padding(.top, -5).font(.system(size: 20.0)).foregroundColor(Color.black)
                     }
+                }
             }
         }
-        .navigationBarTitle("View Media")
-        .navigationBarTitleDisplayMode(.inline)
-        .background(Color.white)
+            .navigationBarTitle("View Media")
+            .navigationBarTitleDisplayMode(.inline)
+            .background(Color.white)
         Spacer()
+    }
+}
+
+struct ConfirmView: View {
+    @Environment(\.managedObjectContext) private var viewContext
+    let item: Media
+    
+    var body: some View {
+        VStack {
+            Text("Are you sure you want to delete?").font(.headline).padding(.top, 20).padding(.bottom, 20)
+            Divider()
+            HStack {
+                Button(action: {
+                    UIApplication.shared.windows[0].rootViewController?.dismiss(animated: true, completion: {})
+                    let _ = viewContext.delete(item)
+                    try? viewContext.save()
+                }) {
+                    Text("Yes").frame(width: 150, height: 30)
+                }.padding(.bottom,5)
+                Divider()
+                Button(action: {
+                    UIApplication.shared.windows[0].rootViewController?.dismiss(animated: true, completion: {})
+                }) {
+                    Text("No").frame(width: 150, height: 30).foregroundColor(Color.red)
+                }.padding(.bottom, 5)
+            }.frame(height: 40).padding(0)
+            }.background(Color(white: 0.9))
     }
 }
 
 struct AlertView: View {
     @Environment(\.managedObjectContext) private var viewContext
     @State private var text: String = ""
-    @State private var showingAlert = false
     let item: Media
     
     var body: some View {
         VStack {
-            Text("Enter Media Name").font(.headline)
-                .padding(.top, 20)
-                .padding(.bottom, 0)
-
-            TextField("Name...", text: $text).textFieldStyle(RoundedBorderTextFieldStyle()).padding()
+            Text("Rename Media").font(.headline).padding(.top, 20).padding(.bottom, 0)
+            TextField("Name...", text: $text)
+                .textFieldStyle(RoundedBorderTextFieldStyle())
+                .padding().padding(.trailing, 10).padding(.leading, 10)
             Divider()
             HStack {
-                Spacer()
                 Button(action: {
                     UIApplication.shared.windows[0].rootViewController?.dismiss(animated: true, completion: {})
             
@@ -117,21 +150,16 @@ struct AlertView: View {
                         }
                     }
                 }) {
-                    Text("Save")
-                }.padding(.bottom,5)
-                Spacer()
+                    Text("Save").frame(width: 240, height: 30)
+                }.padding(.bottom, 5)
                 Divider()
-                Spacer()
                 Button(action: {
                     UIApplication.shared.windows[0].rootViewController?.dismiss(animated: true, completion: {})
                 }) {
-                    Text("Cancel")
+                    Text("Cancel").frame(width: 240, height: 30)
                 }.padding(.bottom,5)
-                Spacer()
-            }
-            .frame(height: 40)
-            .padding(0)
-            }.background(Color(white: 0.9))
+            }.frame(height: 40).padding(0)
+        }.background(Color(white: 0.9))
     }
 }
